create flask app
