+++ site/app/flask/requirements.txt
@@ -1,0 +1,3 @@
+flask
+pymysql
+cryptography
