+++ site/app/flask/static/images/delete.png
