--- site/app/requirements.txt
@@ -1,6 +1,0 @@
-flask
-pymysql
-flask_limiter
-requests
-flask-cors
-cryptography
