+++ site/db/db.env
@@ -1,2 +1,2 @@
-MYSQL_DATABASE=project
+MYSQL_DATABASE=chaimtube
 MYSQL_ROOT_PASSWORD=changeme
