+++ site/docker-compose.yml
@@ -6,18 +6,18 @@ services:
     build: db/.
     env_file:
       - db/db.env
+    networks:
+      - project
   
   app:
     container_name: chaimtube_app
     build: app/.
+    depends_on:
+      - db
     ports:
       - 5000:5000
-
-  # server:
-  #   container_name: chaimtube_server
-  #   build: server/.
-  #   ports:
-  #     - 80:80
+    networks:
+      - project
 
 networks:
   project:
